Refactor favourite_team into teams_utils
